requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
base58==2.1.1
cachetools==6.2.4
certifi==2025.1.31
cffi==2.0.0
charset-normalizer==3.4.1
click==8.3.1
colorama==0.4.6
construct==2.10.70
construct-typing==0.7.0
cryptography==46.0.3
deprecation==2.1.0
distlib==0.3.9
fastapi==0.124.4
filelock==3.18.0
frozenlist==1.8.0
fsspec==2025.10.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
hyperframe==6.1.0
idna==3.10
iniconfig==2.3.0
Jinja2==3.1.6
jsonalias==0.1.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
MetaTrader5==5.0.4874
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.0
networkx==3.5
numpy==2.2.4
ollama==0.6.0
packaging==25.0
pandas==2.2.3
platformdirs==4.3.7
pluggy==1.6.0
postgrest==2.27.2
propcache==0.4.1
protobuf==6.33.0
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
pyiceberg==0.10.0
PyJWT==2.10.1
pyparsing==3.3.1
pyroaring==1.0.3
pytest==9.0.2
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
realtime==2.27.2
regex==2025.11.3
requests==2.32.3
rich==14.2.0
safetensors==0.6.2
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
solana==0.36.10
solders==0.27.1
sortedcontainers==2.4.0
starlette==0.50.0
storage3==2.27.2
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.27.2
supabase-auth==2.27.2
supabase-functions==2.27.2
sympy==1.14.0
ta==0.11.0
tenacity==9.1.2
tiktoken==0.12.0
tokenizers==0.22.1
torch==2.9.0
tqdm==4.67.1
transformers==4.57.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.38.0
virtualenv==20.30.0
watchfiles==1.1.1
websockets==15.0.1
wheel==0.45.1
yarl==1.22.0
